requirements.txt
# Deep Learning Frameworks
torch
torchvision
tensorflow>=2.13.0
keras>=2.13.0

# Computer Vision and Image Processing
opencv-python
Pillow

# Data Processing
numpy
pandas
duckdb
datasets
huggingface_hub

# Machine Learning
scikit-learn

# Visualization
matplotlib
seaborn

# NLP/Transformers (for scene parsing)
transformers

# Utilities
tqdm
